+++ CHANGELOG.md
@@ -15,6 +15,7 @@ the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 - Added: Squanto basic search
 - Added: Show a new tree breadcrumbs column when in Table filtering/sorting mode. This can be (de)selected as column
 - Added: phpstan baseline to temporary ignore error stack
+- Added: better indices on asset pivot table (Run migrations)
 - Changed: On first publication of visitable model, all links are put online automatically
 
 ## [0.10.20] - 2026-03-17